+++ CHANGELOG.md
@@ -1,12 +1,15 @@
-# TBD  !Change this before committing!
+
+# 1.0.0
+
 - Removed some private methods that accidentally escaped into the public API.
 - Removed dependency on dart:io to enable js support.
 - Updated README with basic usage information.
+- Documented more of the public API.
 - Renamed HeliumClient to HeliumBlockchainClient.
 - Renamed HeliumPoCReceiptsPathElement to HeliumPoCPathElement.
 - Renamed HeliumPoCReceiptsWitness to HeliumPoCWitness.
 - Renamed HeliumPoCReceiptsReceipt to HeliumPoCReceipt.
-- Documented more of the public API.
+- Renamed public methods on HeliumHotspotClient, HeliumOraclePricesClient, and HeliumTransactionsClient.
 
 # 0.1.0